+++ MyLib/BasicDataRight.cs
@@ -12,10 +12,10 @@ namespace MyLib
         public readonly int i;
         public readonly string Si;
         public readonly int s;
-        public readonly float CSz;
+        public readonly decimal CSz;
         public readonly DateTime EXb;
         public readonly DateTime EXe;
-        public readonly float TIs;
+        public readonly decimal TIs;
         public readonly int SEt;
         public readonly string EXc;
         public readonly Boolean HOt;
@@ -35,13 +35,13 @@ namespace MyLib
                     else if (word.StartsWith(checks[2]))
                         s = Int32.Parse(word.Substring(1, word.Length - 1));
                     else if (word.StartsWith(checks[3]))
-                        CSz = float.Parse(word.Substring(3, word.Length - 3), CultureInfo.InvariantCulture);
+                        CSz = decimal.Parse(word.Substring(3, word.Length - 3), CultureInfo.InvariantCulture);
                     else if (word.StartsWith(checks[4]))
                         EXb = DateTime.ParseExact(word.Substring(3, word.Length - 3), "yyyyMMdd", CultureInfo.InvariantCulture);
                     else if (word.StartsWith(checks[5]))
                         EXe = DateTime.ParseExact(word.Substring(3, word.Length - 3), "yyyyMMdd", CultureInfo.InvariantCulture);
                     else if (word.StartsWith(checks[6]))
-                        TIs = float.Parse(word.Substring(3, word.Length - 3), CultureInfo.InvariantCulture);
+                        TIs = decimal.Parse(word.Substring(3, word.Length - 3), CultureInfo.InvariantCulture);
                     else if (word.StartsWith(checks[7]))
                         SEt = Int32.Parse(word.Substring(3, word.Length - 3));
                     else if (word.StartsWith(checks[8]))
